selectors - move LightBox needs to standalone selector
